Added support for multiple accounts
Significant refactoring was required, but it seems to work with assuming roles in multiple accounts now. I also changed the "default_region" field to just "region". The downside is that the "region" field is now required. If a task doesn't specify a region, the exporter will not perform that task